+++ src-tauri/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod parse_log_file;
 pub mod plugins;
-pub mod dp_utils;
+pub mod overlay_server;
